% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suppoRRT.R
\docType{data}
\name{huntRRdata}
\alias{huntRRdata}
\title{Randomized response technique (RRT) questionnaire on illegal bird hunting in Southwest China}
\format{A data frame containing a sample of 189 observations.
Each question yielded a yes (\code{1}) or no (\code{0}) response. The RRT questions (response variables) were as follows:
\itemize{
\item \code{Barbet}: In the past year, have you hunted barbet?
\item \code{Bulbul}: In the past year, have you hunted bulbul?
\item \code{Partridge}: In the past year, have you hunted partridge?
}
In addition, the column \code{Age} presents respondent age as a predictor covariate (standardized to z-scores).}
\usage{
data(huntRRdata)
}
\description{
Individuals in Southwest China were asked whether or not they had hunted barbet (\emph{Psilopogon} spp.), bulbul (\emph{Pycnonotus} spp.), and/or partridge (\emph{Arborophila} spp) using the forced response RRT question design. This dataset presents one covariate: respondent age (centered and scaled).
}
\examples{
## Loading the data
data(huntRRdata)
p00 <- p11 <- 5/6

## Performing univariate RRT inference
   # How many individuals hunted barbets?
RRunivariate(y=huntRRdata['Barbet'], p00=p00, p11=p11)
   # Was age predictive of hunting barbets?
RRunivariate(y=huntRRdata['Barbet'], x=huntRRdata['Age'], p00=p00, p11=p11)

## Using RRbootstrap to determine confidence intervals for "Intercept" and Age
   # NB: Increase reps for more robust analyses
RRbootstrap(model="univariate", y=huntRRdata['Barbet'], x=huntRRdata['Age'],
            p00=p00, p11=p11, reps=50)

## Performing multivariate inference
    # Sum score model
RRsumscore(y=huntRRdata[,1:3], p00=p00, p11=p11)
    # Item response theory model
RRirt(y=huntRRdata[,1:3],p00=p00, p11=p11)

}
\references{
Chang, C.H., Williams, S.J., Cruyff, M.J.L.F., Zhang, M., Wilcove, D.S., Levin, S.A., and Quan, R.-C. (2017).
\emph{Predictors of illegal hunting in Southwest China.}
Unpublished Data.
}
\seealso{
\code{RRdata} for RRT data on sensitive behaviors among university students.
}
\keyword{datasets}
