% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suppoRRT.R
\name{RRbootstrap}
\alias{RRbootstrap}
\title{Bootstrap estimates for RRT models}
\usage{
RRbootstrap(model = c("univariate", "sumscore", "irt"), y, x = NULL, p00,
  p11, noncompliance = F, reps = 500, seed = 1)
}
\arguments{
\item{model}{A character string, one of \code{"univariate","sumscore",} or \code{"irt"}.}

\item{y}{Matrix or data frame of observed responses.}

\item{x}{Predictor variable values (matrix or data frame).}

\item{p00}{The conditional classification probability \eqn{y=0|z=0} (observed no conditional on true no).}

\item{p11}{The conditional classification probability \eqn{y=1|z=1} (observed yes conditional on true yes).}

\item{noncompliance}{Should the model specify question noncompliance (one-sided lying to avoid saying "yes")? \code{FALSE} (default) or \code{TRUE}. Noncompliance is only included in the \code{irt} model.}

\item{reps}{The number of samples with replacement that \code{RRbootstrap} takes to generate the estimates.}

\item{seed}{See \code{\link[base]{Random}}.}
}
\value{
\code{RRbootstrap} returns a list \code{$bootresult} that provides the mean, 50\%, 2.5\%, and 97.5\% quantile bootstrap estimates for parameters generated by each RRT model.
}
\description{
\code{RRbootstrap} generates bootstrap estimates for univariate (\code{\link{RRunivariate}}) and multivariate (\code{\link{RRsumscore}} and \code{\link{RRirt}}) RRT models, both with and without covariates. \code{RRbootstrap} is intended to allow users to deal with issues arising from non-invertible Hessians where standard errors cannot otherwise be computed. We caution, however, that when standard errors cannot be computed, that often gestures to issues with the underlying data (too few observations or traits that are too rare) and/or poorly-specified models.
}
\examples{
## BOOTSTRAP ESTIMATES FOR ITEM RESPONSE THEORY
# Step 1: Load or simulate multivariate RRT data
pi.1s <- c(0.15, 0.2, 0.3)
p00 <- p11 <- 5/6
IRTdata <- RRsimulate(model="irt", pi.1 = pi.1s, p00=p00, p11=p11, x.betas=c(-1, 3), n=200)

# Step 2: Run RRbootstrap (NB: increase reps for actual analyses)
IRT.boot <- RRbootstrap("irt", y=IRTdata[,1:3], p00=p00, p11=p11, reps=50)
# Not run:
# IRT.boot <- RRbootstrap("irt", y=IRTdata[,1:3], x=IRTdata[,4:5], p00=p00, p11=p11, reps=500)
}
\seealso{
\code{\link{RRunivariate}} for single-item RRT estimation, \code{\link{RRsumscore}} for sum score modeling, and \code{\link{RRirt}} for item randomized response theory modeling. The latter two models (sum score and IRT) are suitable for multivariate RRT data (two or more question prompts).
}
