% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suppoRRT.R
\name{RRgof}
\alias{RRgof}
\title{Goodness of Fit Calculation for RRT models}
\usage{
RRgof(model = c("univariate", "sumscore", "irt"), y, x = NULL, p00, p11,
  noncompliance = F, reps = 500, seed = 1)
}
\arguments{
\item{model}{A character string, one of \code{"univariate","sumscore",} or \code{"irt"}.}

\item{y}{Matrix or data frame of observed responses.}

\item{x}{Predictor variable values (matrix or data frame).}

\item{p00}{The conditional classification probability \eqn{y=0|z=0} (observed no conditional on true no).}

\item{p11}{The conditional classification probability \eqn{y=1|z=1} (observed yes conditional on true yes).}

\item{noncompliance}{Should the model specify question noncompliance (one-sided lying to avoid saying "yes")? \code{FALSE} (default) or \code{TRUE}. Noncompliance is only included in the \code{irt} model (see \code{\link{RRirt}} for more information).}

\item{reps}{The number of parametric bootstraps performed to assess goodness of fit (default=500)}

\item{seed}{See \code{\link[base]{Random}}.}
}
\value{
\code{RRbootstrap} returns a list. \code{$gof.test.result} provides a data frame with the model string, \eqn{\chi^2_0} value (\code{chiSquare0}), degrees of freedom (\code{df}), p-value observed for original data against the model fit (\code{chiSquarePval}), and the parametric bootstrap p-value generated from resampling observations based on the original model (\code{boot.p.value}). \code{$X2boot} provides the bootstrapped \eqn{\chi^2_{boot}} values.
}
\description{
\code{RRgof} calculates a parametric bootstrap \eqn{\chi^2} statistic for the goodness-of-fit (MacKenzie and Bailey 2004) of the three RRT models presented in this package: univariate (\code{\link{RRunivariate}}) and multivariate (\code{\link{RRsumscore}} and \code{\link{RRirt}}) RRT models, both with and without covariates. \code{RRgof} should be used to evaluate the model fit generated by one of the RRT models (\code{\link{RRunivariate}}, \code{\link{RRsumscore}}, or \code{\link{RRirt}}); it does not return parameter estimates for the models.
}
\examples{
## GOODNESS OF FIT ESTIMATES
### UNIVARIATE MODEL EXAMPLES
data(huntRRdata)
p00 <- p11 <- 5/6
reps <- 100
gof.univariate <- RRgof(model="univariate", y=huntRRdata['Bulbul'], p00=p00, p11=p11, reps=reps)
gof.univariate$gof.test.result
gof.univariate.GLM <- RRgof(model="univariate", y=huntRRdata['Bulbul'], x=huntRRdata['Age'], p00=p00, p11=p11, reps=reps)
gof.univariate.GLM$gof.test.result
### IRT MODEL EXAMPLES (with noncompliance)
data(huntRRdata)
p00 <- p11 <- 5/6
reps <- 10 # Increase the number of reps for your own calculations (10 chosen for convenience)
gof.IRT.GLM <- RRgof(model="irt", y=huntRRdata[,c(1:3)], x=huntRRdata['Age'], noncompliance=TRUE, p00=p00, p11=p11, reps=reps)
}
\references{
MacKenzie, Darryl I and Bailey, Larissa L. (2004).
\emph{Assessing the fit of site-occupancy models.}
 Journal of Agricultural, Biological, and Environmental Statistics. 9(3):300-318.
}
\seealso{
\code{\link{RRunivariate}} for single-item RRT estimation, \code{\link{RRsumscore}} for sum score modeling, and \code{\link{RRirt}} for item randomized response theory modeling. The latter two models (sum score and IRT) are suitable for multivariate RRT data (two or more question prompts).
}
