% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suppoRRT.R
\name{RRsimulate}
\alias{RRsimulate}
\title{Simulate RRT data for three models: univariate, sum score, and item response theory}
\usage{
RRsimulate(model = c("univariate", "sumscore", "irt"), pi.1 = c(0.2),
  p00 = 5/6, p11 = 5/6, n = 100, x.betas = c(0), seed = 1)
}
\arguments{
\item{model}{A character string, one of \code{"univariate","sumscore",} or \code{"irt"}.}

\item{pi.1}{A numeric vector with the true prevalence of sensitive trait(s) (each value must be within \code{c(0,1)}).}

\item{p00}{Probability of answering \code{no} conditional on a true \code{no} (must be within c(0,1))}

\item{p11}{Probability of answering \code{yes} conditional on a true \code{yes}}

\item{n}{Sample size (number of respondents)}

\item{x.betas}{A numeric vector of predictor variable regression coefficients (defaults to \code{c(0)})}

\item{seed}{See \code{\link[base]{Random}}.}
}
\value{
\code{RRsimulate} returns a data frame with y (RRT responses) and x (predictor variable) observations.
}
\description{
\code{RRsimulate} simulates univariate (single item) and multivariate (multiple items) RRT data. Probabilities (\code{pi.1, p00, p11}) must be within the interval (0, 1).
}
\examples{
## UNIVARIATE SIMULATION - with no informative predictor variable
UNIdata <- RRsimulate(model="univariate", pi.1 = 0.35, p00=5/6, p11=5/6, n=200)
       ## Simulate with an informative predictor
UNIdataLR <- RRsimulate(model="univariate",pi.1 = 0.35, p00=5/6, p11=5/6, x.betas=c(2), n=200)
head(UNIdataLR)

## SUM SCORE SIMULATION
pi.1s <- c(0.2, 0.35, 0.4)
p00 <- p11 <- 5/6
SSdata <- RRsimulate(model="sumscore",pi.1 = pi.1s, p00=p00, p11=p11, x.betas=c(-1, 3), n=200)
head(SSdata)

## ITEM RESPONSE THEORY SIMULATION
pi.1s <- c(0.15, 0.2, 0.3)
p00 <- p11 <- 5/6
IRTdata <- RRsimulate(model="irt", pi.1 = pi.1s, p00=p00, p11=p11, x.betas=c(-1, 3), n=200)
head(IRTdata)

}
\seealso{
\code{\link{RRunivariate}} for single-item RRT estimation, \code{\link{RRsumscore}} for sum score modeling, and \code{\link{RRirt}} for item randomized response theory modeling. The latter two models (sum score and IRT) are suitable for multivariate RRT data (two or more question prompts).
}
