% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RRunivariate.R
\name{RRunivariate}
\alias{RRunivariate}
\title{Single-item (univariate) RRT modeling for prevalence and logistic regression}
\usage{
RRunivariate(y, x = NULL, p00, p11, n.round = 3, bootin = FALSE)
}
\arguments{
\item{y}{Observed responses (vector or 1-column matrix/data frame)}

\item{x}{Predictor variable values (data frame). An intercept is automatically added.}

\item{p00}{The conditional classification probability \eqn{y=0|z=0} (observed no conditional on true no)}

\item{p11}{The conditional classification probability \eqn{y=1|z=1} (observed yes conditional on true yes)}

\item{n.round}{Number of digits for rounding estimates (default is \code{n.round=3})}

\item{bootin}{Default: \code{bootin=FALSE} - do not change. For RRbootstrap internal functioning.}
}
\value{
\code{RRunivariate} returns a list:
\itemize{
   \item \code{$estimates}: a data frame with the parameter estimate (\code{est}), standard error (\code{se}), T-statistic (\code{tval}), lower bounds of the 95\% confidence interval (\code{min95}), and upper bounds of the 95\% confidence interval (\code{max95}).
   \item \code{$fitmeasures}: -2*log likelihood (\code{minus2LL}), degrees of freedom (df), Akaike Information Criterion (\code{AIC}), and AIC corrected for (small) sample sizes (\code{AICc}).
}
When the Hessian is singular, standard errors are inestimable, as are the t-statistic, p-value, and confidence limits; in this case, the model will return \code{NaN} for those entries.
}
\description{
\code{RRunivariate} finds (1) the prevalence of a sensitive trait given a specific RRT question design, and (2) can perform logistic regression for individual RRT items.
}
\examples{
## Univariate analysis of simulated data (see ?RRsimulate for questions)
RRTdata <- RRsimulate(pi.1=0.3, p00=5/6, p11=5/6, n=300,
                      model="univariate", x.betas=c(-1,3))
head(RRTdata) # see first 6 rows of simulated data
   # Univariate estimation of sensitive trait prevalence
RRunivariate(RRTdata[,1], p00=5/6, p11=5/6)
   # Univariate logistic regression
RRunivariate(y=RRTdata[,1], x=RRTdata[,2:3], p00=5/6, p11=5/6)

## Univariate analysis of bulbul hunting in Southwest China
    # Load the data and prepare the design parameters
data(huntRRdata) # See ?huntRRdata for information on this dataset
p00 <- p11 <- 5/6
    # Prevalence of hunting bulbuls
RRunivariate(y=huntRRdata['Bulbul'], p00=p00, p11=p11)
    # Effect of age on the probability of hunting bulbuls
RRunivariate(y=huntRRdata['Bulbul'], x=huntRRdata['Age'],p00=p00, p11=p11)

## Univariate analysis of sensitive behaviors among university students
   # Loading the data
data(RRdata) # See ?RRdata for more information on the sensitive questionnaire.

   # Calculating p00 and p11 for each question
college.py <- c(1/12, 1/10, 20/30, 1/10, 10/30, 1/12)
college.p00 <- 1-0.5*college.py # p00 for each question prompt in order
college.p11 <- 1-0.5*(1-college.py) # p11 for each question prompt in order

   # Performing univariate RRT inference
     # How many students copied exams? (Question 1)
RRunivariate(y=RRdata['copied'], p00=college.p00[1], p11=college.p11[1])
     # How many students took drugs on campus? (Question 5)
RRunivariate(RRdata['drug'], p00=college.p00[5], p11=college.p11[5])

}
\seealso{
\code{\link{RRbootstrap}} for bootstrap estimates when the Hessian is singular.
}
