% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suppoRRT.R
\docType{package}
\name{zapstRR}
\alias{zapstRR}
\alias{zapstRR-package}
\alias{zapstRR-package}
\title{zapstRR: A package for performing statistical inference on univariate and multivariate Randomized Response Technique (RRT) data}
\description{
zapstRR provides three categories of functions: RRT analysis, power analysis, and data simulation.
}
\details{
\tabular{ll}{
Package: \tab zapstRR\cr
Type: \tab Package\cr
Version: \tab 0.0.0.900\cr
Date: \tab 2017-10-10\cr
Depends: \tab R (>= 3.3.0)\cr
Imports: \tab stats, grDevices, graphics\cr
Suggests: \tab knitr\cr
License: \tab GPL (>= 2)\cr
LazyLoad: \tab yes\cr
URL: \tab \url{https://github.com/charlottehchang/zapstRR}\cr
Vignette: \tab \url{http://rpubs.com/chwchang/318110}\cr
}
}
\section{Package Functions}{

The RRT analysis functions are as follows:
\itemize{
     \item \code{\link{RRunivariate}}: Prevalence and logistic regression for single RRT items
     \item \code{\link{RRsumscore}}: Prevalence and ordinal regression for sum scores of RRT items
     \item \code{\link{RRirt}}: Prevalence and multivariate regression for two or more RRT items using item response theory
     \item \code{\link{RRbootstrap}}: Nonparametric bootstrap estimates of confidence intervals for the three models above.
}
In addition, \code{\link{powerRRT}} performs power analysis for study designs and \code{\link{RRsimulate}} can simulate example RRT data. Two datasets are provided: \code{\link{huntRRdata}}, a case study on illegal hunting in Southwest China, and \code{\link{RRdata}}, which investigates sensitive behaviors among university students.
}

